Struct ethers_middleware::gas_escalator::GasEscalatorMiddleware

pub struct GasEscalatorMiddleware<M, E> {
    pub txs: Arc<Mutex<Vec<(TxHash, TransactionRequest, Instant, Option<BlockId>)>>>,
    // some fields omitted
}

A Gas escalator allows bumping transactions’ gas price to avoid getting them stuck in the memory pool.

use ethers_providers::{Provider, Http};
use ethers_middleware::{
    gas_escalator::{GeometricGasPrice, Frequency, GasEscalatorMiddleware},
    gas_oracle::{EthGasStation, GasCategory, GasOracleMiddleware},
};
use std::{convert::TryFrom, time::Duration, sync::Arc};

let provider = Provider::try_from("http://localhost:8545")
    .unwrap()
    .interval(Duration::from_millis(2000u64));

let provider = {
    let escalator = GeometricGasPrice::new(5.0, 10u64, None::<u64>);
    GasEscalatorMiddleware::new(provider, escalator, Frequency::PerBlock)
};

// ... proceed to wrap it in other middleware
let gas_oracle = EthGasStation::new(None).category(GasCategory::SafeLow);
let provider = GasOracleMiddleware::new(provider, gas_oracle);

Fields
txs: Arc<Mutex<Vec<(TxHash, TransactionRequest, Instant, Option<BlockId>)>>>

The transactions which are currently being monitored for escalation

Implementations

impl<M, E> GasEscalatorMiddleware<M, E> where
M: Middleware,
E: GasEscalator,

pub fn new(inner: M, escalator: E, frequency: Frequency) -> Self where
E: Clone + 'static,
M: Clone + 'static,

Initializes the middleware with the provided gas escalator and the chosen escalation frequency (per block or per second)

pub async fn escalate(&self) -> Result<(), GasEscalatorError<M>>

Re-broadcasts pending transactions with a gas price escalator