Module ethers_middleware::gas_oracle

The gas oracle middleware is used to get the gas price from a list of gas oracles instead of using eth_gasPrice. For usage examples, refer to the GasOracle trait.

Re-exports

pub use blocknative::BlockNative;

pub use eth_gas_station::EthGasStation;

Deprecated

pub use etherchain::Etherchain;

pub use etherscan::Etherscan;

pub use middleware::GasOracleMiddleware;

pub use middleware::MiddlewareError;

pub use median::Median;

pub use cache::Cache;

pub use polygon::Polygon;

pub use gas_now::GasNow;

pub use provider_oracle::ProviderOracle;

Modules

blocknative

cache

eth_gas_station

etherchain

etherscan

gas_now

median

middleware

polygon

provider_oracle

Enums

GasCategory

Generic GasOracle gas price categories.

GasOracleError

Error thrown by a GasOracle.

Traits

GasOracle

An Ethereum gas price oracle.

Type Aliases

Result