ethers_middleware::transformer
Trait Transformer

pub trait Transformer:
    Send
    + Sync
    + Debug {
    // Required method
    fn transform(
        &self,
        tx: &mut TypedTransaction,
    ) -> Result<(), TransformerError>;
}

Transformer is a trait to be implemented by a proxy wallet, eg. DsProxy, that intends to intercept a transaction request and transform it into one that is instead sent via the proxy contract.

Required Methods

fn transform(&self, tx: &mut TypedTransaction) -> Result<(), TransformerError>

Transforms a transaction request into one that can be broadcasted and execute via the proxy contract.